streamlit_app/requirements.txt
streamlit
psycopg2-binary
python-dotenv
pm4py        # for later steps
pandas       # for later steps
